% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.viz.R
\name{ds.viz}
\alias{ds.viz}
\title{Create visualizations from DataShield data}
\usage{
ds.viz(connections, variable, breaks = 10)
}
\arguments{
\item{connections}{DataShield connections}

\item{variable}{variable name to analyze}

\item{breaks}{number of histogram breaks}
}
\value{
summary statistics data frame
}
\description{
Creates histogram and boxplot from aggregated data
}
\author{
Your Name
}
